+++ src/all/data/js/quickaccess/popup/components/HomePage/canSuggestUrl.js
@@ -2,53 +2,110 @@ import XRegExp from "xregexp";
 import ipRegex from "ip-regex";
 
 /**
- * Check if an url can be suggested for a given hostname.
- * An url can be suggested, if:
- * - It matches the hostname;
- * - It is a parent host of the hostname;
+ * Check if an url can be suggested for a given one.
+ * The urls can be suggested if:
+ * - The urls hostname match;
+ * - The suggested url is a parent of the url;
+ * - (optional check) If the suggested url contains a protocol, then it has to match the url.
+ * - (optional check) If the suggested url contains a port, then it has to match the url.
  *
- * @param {string} hostname The domain
- * @param {resource} resource The resource
+ * Note that this function does not take in account any urls path, parameters or hash.
+ *
+ * @param {string} url The url.
+ * @param {string} suggestedUrl The potential url to suggest
  * @returns {boolean}
  */
-export default function (hostname, url) {
-  const urlHostname = extractUrlHostname(url);
+export default function (url, suggestedUrl) {
+  let urlObject;
+  let suggestedUrlObject;
 
-  if (!urlHostname) {
+  try {
+    urlObject = new URL(url);
+  } catch (error) {
+    // Only valid url are accepted by this function.
+    // This information should come from the browser tab url, and so should be valid.
     return false;
   }
 
-  if (hostname === urlHostname) {
-    return true;
+  suggestedUrlObject = parseSuggestedUrl(suggestedUrl);
+
+  // Unable to parse the suggested url.
+  if (!suggestedUrlObject) {
+    return false;
   }
 
-  // If IPs, return a strict comparison.
-  const hostnameIsIpAddress = ipRegex({exact: true}).test(hostname);
-  const urlHostnameIsIpAddress = ipRegex({exact: true}).test(urlHostname);
-  if (hostnameIsIpAddress || urlHostnameIsIpAddress) {
-    return hostname === urlHostname;
-  }
-
-  // Otherwise check if the url hostname is a parent host of the given hostname.
-  return isParentHostname(urlHostname, hostname);
-}
-
-/**
- * Extract the hostname of an url.
- * @param {string} url The url to work with.
- * @returns {string|boolean} Return the url hostname, false in case of error.
- */
-const extractUrlHostname = function(url) {
-  try {
-    return (new URL(url)).hostname;
-  } catch (error) {
-    // URL tool works only with absolute URL, retry with a forged absolute url.
-    try {
-      return (new URL(`https://${url}`)).hostname;
-    } catch (error) {
+  // Check the protocol, if the suggest url has defined it.
+  if (suggestedUrlObject.protocol) {
+    if (urlObject.protocol !== suggestedUrlObject.protocol) {
       return false;
     }
   }
+
+  // Check the port, if the suggest url has defined it.
+  if (suggestedUrlObject.port) {
+    if (urlObject.port !== suggestedUrlObject.port) {
+      return false;
+    }
+  }
+
+  // Perfect match.
+  if (urlObject.hostname === suggestedUrlObject.hostname) {
+    return true;
+  }
+
+  // If IPs, make a strict comparison.
+  const urlIsIpAddress = ipRegex({exact: true}).test(urlObject.hostname);
+  const suggestUrlIsIpAddress = ipRegex({exact: true}).test(suggestedUrlObject.hostname);
+  if (urlIsIpAddress || suggestUrlIsIpAddress) {
+    return urlObject.hostname === suggestedUrlObject.hostname;
+  }
+
+  // Otherwise check if the suggested url hostname is a parent host of the url hostname.
+  return isParentHostname(suggestedUrlObject.hostname, urlObject.hostname);
+}
+
+/**
+ * Parse a suggested url.
+ * @param {string} suggestedUrl The url.
+ * @returns {object}
+ * {
+ *   protocol: {string},
+ *   hostname: {string},
+ *   port: {integer}
+ * }
+ */
+const parseSuggestedUrl = function (suggestedUrl) {
+  let suggestedUrlObject;
+  let protocol = "";
+  let hostname = "";
+  let port = "";
+  let enforceProtocol = false;
+
+  // The browser URL primitive does not work with relative urls.
+  // Enforce a protocol to the suggested url, if none has been defined. A fake protocol is used in order to preserve
+  // the port information that can be altered by the parsing. By instance, when the https is enforced to the url
+  // www.passbolt.com:443, then the port information is deleted after the parsing.
+  if (!/^[a-zA-Z\-]*:\/\//.test(suggestedUrl)) {
+    enforceProtocol = true;
+    suggestedUrl = `fake://${suggestedUrl}`;
+  }
+
+  try {
+    suggestedUrlObject = new URL(suggestedUrl);
+  } catch (error) {
+    return false;
+  }
+
+  port = suggestedUrlObject.port;
+  if (!enforceProtocol) {
+    protocol = suggestedUrlObject.protocol;
+    hostname = suggestedUrlObject.hostname;
+  } else {
+    suggestedUrlObject.protocol = "https:";
+    hostname = suggestedUrlObject.hostname;
+  }
+
+  return {protocol, hostname, port};
 };
 
 // Hostname allowed characters regex